Refactor code by breaking UI into logical parts.
