+++ project/build.sbt
@@ -7,6 +7,8 @@ version := "1.0-SNAPSHOT"
 
 scalaVersion := "2.11.6"
 
+scalacOptions += "-Xexperimental"
+
 libraryDependencies ++= Seq(
     "scalation" % "scalation_2.11" % "1.1",
     "org.scalatest" % "scalatest_2.11" % "2.2.1" % "test",
